deepen: FileView owns read path (inline file-reader/read-render/truncate)
